Code gedeeltelijk opgeruimd en verduidelijkt.
