+++ templates/home.html
@@ -26,15 +26,6 @@
         </nav>
     </header>
     <main id="body">
-        <h2> De eerste game </h2>
-        <div>
-            <p><strong>Naam:</strong> {{ eerste_game['name'] }} <br></p>
-            <p><strong>Prijs:</strong> &euro; {{ eerste_game['price'] }}<br></p>
-            <p><strong>Genres:</strong> {{ eerste_game['genres'] }}</p>
-        </div>
-
-        <img src="static/images/graph_price.png" alt="Graph for price categories">
-        <img src="static/images/graph_genre.png" alt="">
         <img src="static/images/graph_rapportcijfers.png" alt="">
         <img src="static/images/graph_ratings_20.png" alt="">
     </main>
